Training and Testing Data
The files: [titanic_test.csv] and [titanic_train.csv] are used for training and testing the models. These files have been downloaded from the Kaggle Repository for Titanic Crash survivors and can be found there.
Both the files contain similar data that states information about all the passengers that were travelling on the Titanic. It has features like the class of cabin, physical attributes, gender, fare they paid; and based out of all this information, the prediction is made: Given a set of features, whether a traveler would have survived.